refactor: 패키지화
