--- src/main/java/domain/Name.java
@@ -1,32 +1,0 @@
-package domain;
-
-import java.util.Objects;
-
-public class Name {
-    private final String name;
-
-    public Name(final String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        final Name name1 = (Name) o;
-        return Objects.equals(name, name1.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
-    }
-}
