--- src/main/java/domain/Participant.java
@@ -1,14 +1,0 @@
-package domain;
-
-public class Participant extends Player {
-    private static final int BUST_CONDITION = 21;
-
-    public Participant(final Name name) {
-        super(name);
-    }
-
-    @Override
-    public boolean isNotBust() {
-        return calculateScore() < BUST_CONDITION;
-    }
-}
