+++ src/main/java/domain/card/Card.java
@@ -1,25 +1,42 @@
 package domain.card;
 
-import dto.CardResponse;
+import java.util.Objects;
 
 public class Card {
-    private final Rank rank;
-    private final Suit suit;
+    private final Denomination denomination;
+    private final Symbol symbol;
 
-    public Card(final Rank rank, final Suit suit) {
-        this.rank = rank;
-        this.suit = suit;
+    public Card(final Denomination number, final Symbol symbol) {
+        this.denomination = number;
+        this.symbol = symbol;
     }
 
-    public CardResponse toCardResponse() {
-        return new CardResponse(rank.name(), suit.name());
+    public int getValue() {
+        return denomination.getValue();
     }
 
-    public int getRankValue() {
-        return rank.getValue();
+    public Denomination getDenomination() {
+        return denomination;
     }
 
-    public Rank getRank() {
-        return rank;
+    public Symbol getSymbol() {
+        return symbol;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        final Card card = (Card) o;
+        return denomination == card.denomination && Objects.equals(symbol, card.symbol);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(denomination, symbol);
     }
 }
