+++ src/main/java/domain/card/Symbol.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum Symbol {
     HEART,
