+++ src/main/java/domain/player/Name.java
@@ -3,34 +3,16 @@ package domain.player;
 import java.util.Objects;
 
 public class Name {
-    private static final int MAX_LENGTH = 5;
-    private final String value;
+    public static final Name DEALER_NAME = new Name("딜러");
+
+    private final String name;
 
     public Name(final String name) {
-        validate(name);
-        this.value = name;
+        this.name = name;
     }
 
-
-    private void validate(final String name) {
-        validateEmptiness(name);
-        validateLength(name);
-    }
-
-    private void validateLength(final String name) {
-        if (name.length() > MAX_LENGTH) {
-            throw new IllegalArgumentException(String.format("이름 길이는 %d자 이하로 입력해주세요", MAX_LENGTH));
-        }
-    }
-
-    private void validateEmptiness(final String name) {
-        if (name == null || name.isBlank()) {
-            throw new IllegalArgumentException("이름은 비어있을 수 없습니다.");
-        }
-    }
-
-    public String getValue() {
-        return value;
+    public String getName() {
+        return name;
     }
 
     @Override
@@ -42,11 +24,11 @@ public class Name {
             return false;
         }
         final Name name1 = (Name) o;
-        return Objects.equals(value, name1.value);
+        return Objects.equals(name, name1.name);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(value);
+        return Objects.hash(name);
     }
 }
