+++ src/test/java/domain/DeckTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.card.Card;
+import domain.card.Deck;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
